HBSD: Make security.bsd.unprivileged_proc_debug per-jail
Teach HardenedBSD to set security.bsd.unprivileged_proc_debug per-jail.
This is needed to create jails in which the Address Sanitizer (ASAN)
fully works as ASAN utilizes libkvm to inspect the virtual address
space. Instead of having to allow unprivileged process debugging for the
entire system, allow setting it on a per-jail basis.

The sysctl node is still security.bsd.unprivileged_proc_debug and the
jail(8) param is allow.unprivileged_proc_debug. The sysctl code is now a
sysctl proc rather than a sysctl int. This allows us to determine
setting the flag for the corresponding jail (or prison0).

As part of the change, the dynamic allow.* API needed to be modified to
take into account pr_allow flags which may now be disabled in prison0.
This prevents conflicts with new pr_allow flags (like that of vmm(4))
that are added (and removed) dynamically.

Also teach the jail creation KPI to allow differences for certain
pr_allow flags between the parent and child jail. This can happen when
unprivileged process debugging is disabled in the parent prison, but
enabled in the child.

This patch is a candidate for upstreaming to FreeBSD.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Submitted-by:	G2, Inc
github-issue:	#360
MFC-to:		12-STABLE