HBSD: Resolve merge conflicts
This commit simply resolves the merge conflicts between FreeBSD's ASR
implementation and HardenedBSD's ASLR implementation. It boots
successfully. However, work needs to be done to make the two
implementations play well with each other. When one implementation is
enabled, the other must be disabled. Simply resolving the merge conflict
is not enough.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>