HBSD: Re-divorce RETPOLINE logic between base and ports
HardenedBSD differentiates between options meant for base and for ports.
As such, we need to ensure that ${MK_RETPOLINE} is defined prior to
referencing it in `bsd.prog.mk` just as we do in `bsd.lib.mk`.

Note that this is a direct commit to hardened/12-stable/master due to
this logic already being applied in hardened/current/master.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Reported-by:	Marin Bernard