HBSD: Fix FreeBSD's ASR logic when mixed with HardenedBSD's ASLR
When both FreeBSD's ASR and PIE randomization are enabled
(kern.elf64.aslr.enable and kern.elf64.aslr.pie_enable respectively),
the logic for setting the map's ASLR flag was skipped. This was likely
due to an improper merge conflict resolution on HardenedBSD's side.

This commit changes the conditional to use a logical OR rather than a
logicial AND in order to fix the determination to randomize the PIE
base.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Reported-by:	@flyingplastic
issue:		#41
MFC-to:		13-STABLE
MFC-to:		12-STABLE
(cherry picked from commit 25cf93ee4c1b4d33be7ddfb7cef538246b704867)
Signed-off-by: Shawn Webb <shawn.webb@hardenedbsd.org>