HBSD: Disable Cross-DSO CFI for a few core applications/libraries.
Early on in the boot process, the core ZFS utilities crash. I need to be
able to set up a chroot to test individual cases. If I can't boot, then
I can't test.

The goal is to re-enable Cross-DSO CFI for (nearly) every part of the
base OS. We're not there, yet.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>