HBSD: Fix typo
This typo was the cause of the segfault issues reported on the users
mailing list. The conditional was incorrect, causing the VDSO's
randomization not to be reflected in certain cases.

Note that this was only an issue on 13-STABLE due to a bad merge
conflict resolution. 14-CURRENT (the hardened/current/master branch) has
the correct conditional.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Reported-by:	Carlos Lopez