HBSD: Incorporate select changes from a French firewall vendor
Though this commit isn't tested, this brings in some select changes from
a French firewall vendor. In this commit, I've made their changes
generic such that anyone basing their work on this can do so with
minimal changes downstream.

I'm 100% sure further refinement is necessary. However, this provides a
good starting place for further testing and refinement. As we (at
${DAYJOB}, the sponsor of this work) work on integration with the other
repositories, we'll likely find things to clean up or bugs to fix.

This is the first commit in the hardened/13-stable/blackhawk feature
branch. Forward-porting to 14-CURRENT and eventual integration of a few
things into the respective master branches (currently:
hardened/current/master and hardened/13-stable/master) will happen once
we work out those finer details. An example of what could be merged into
master is the pkgbase work.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Sponsored-by:	BlackhawkNest, Inc